Extract Consensus-related functions from Chain struct into a separate file
